+++ e2e-tests/development-runtime/src/components/static-query/use-static-query/index.js
@@ -1,0 +1,5 @@
+export { default as Inline } from "./inline-query"
+export { default as Variable } from "./variable-query"
+export { default as ExportedVariable } from "./exported-variable-query"
+export { default as Destructuring } from "./destructuring"
+export { default as Hot } from "./hot"