+++ packages/gatsby-link/src/__tests__/index.js
@@ -74,6 +74,13 @@ describe(`<Link />`, () => {
     expect(container).toMatchSnapshot()
   })
 
+  it(`matches partially active snapshot`, () => {
+    const { container } = setup({
+      linkProps: { to: `/active/nested`, partiallyActive: true },
+    })
+    expect(container).toMatchSnapshot()
+  })
+
   it(`does not fail to initialize without --prefix-paths`, () => {
     expect(() => {
       getInstance({})
@@ -176,5 +183,12 @@ describe(`ref forwarding`, () => {
 
     expect(innerRef).toHaveBeenCalledTimes(1)
     expect(innerRef).toHaveBeenCalledWith(expect.any(HTMLElement))
+  })
+
+  it(`handles a RefObject (React >=16.4)`, () => {
+    const ref = React.createRef(null)
+    setup({ linkProps: { ref } })
+
+    expect(ref.current).toEqual(expect.any(HTMLElement))
   })
 })