+++ packages/gatsby-source-wordpress/src/normalize.js
@@ -2,6 +2,7 @@ const crypto = require(`crypto`)
 const deepMapKeys = require(`deep-map-keys`)
 const _ = require(`lodash`)
 const { createRemoteFileNode } = require(`gatsby-source-filesystem`)
+const { URL } = require(`url`)
 
 const colorized = require(`./output-color`)
 const conflictFieldPrefix = `wordpress_`
@@ -178,12 +179,14 @@ exports.excludeUnknownEntities = entities =>
 // Create node ID from known entities
 // excludeUnknownEntities whitelisted types don't contain a wordpress_id
 // we create the node ID based upon type if the wordpress_id doesn't exist
-exports.createGatsbyIds = (createNodeId, entities) =>
+exports.createGatsbyIds = (createNodeId, entities, _siteURL) =>
   entities.map(e => {
     if (e.wordpress_id) {
-      e.id = createNodeId(`${e.__type}-${e.wordpress_id.toString()}`)
+      e.id = createNodeId(
+        `${e.__type}-${e.wordpress_id.toString()}-${_siteURL}`
+      )
     } else {
-      e.id = createNodeId(e.__type)
+      e.id = createNodeId(`${e.__type}-${_siteURL}`)
     }
     return e
   })
@@ -230,8 +233,10 @@ exports.mapAuthorsToUsers = entities => {
 }
 
 exports.mapPostsToTagsCategories = entities => {
-  const tags = entities.filter(e => e.__type === `wordpress__TAG`)
-  const categories = entities.filter(e => e.__type === `wordpress__CATEGORY`)
+  const categoryTypes = [`wordpress__wc_categories`, `wordpress__CATEGORY`]
+  const tagTypes = [`wordpress__TAG`, `wordpress__wc_tags`]
+  const tags = entities.filter(e => tagTypes.includes(e.__type))
+  const categories = entities.filter(e => categoryTypes.includes(e.__type))
 
   return entities.map(e => {
     // Replace tags & categories with links to their nodes.
@@ -239,7 +244,11 @@ exports.mapPostsToTagsCategories = entities => {
     let entityHasTags = e.tags && Array.isArray(e.tags) && e.tags.length
     if (tags.length && entityHasTags) {
       e.tags___NODE = e.tags.map(
-        t => tags.find(tObj => t === tObj.wordpress_id).id
+        t =>
+          tags.find(
+            tObj =>
+              (Number.isInteger(t) ? t : t.wordpress_id) === tObj.wordpress_id
+          ).id
       )
       delete e.tags
     }
@@ -248,7 +257,11 @@ exports.mapPostsToTagsCategories = entities => {
       e.categories && Array.isArray(e.categories) && e.categories.length
     if (categories.length && entityHasCategories) {
       e.categories___NODE = e.categories.map(
-        c => categories.find(cObj => c === cObj.wordpress_id).id
+        c =>
+          categories.find(
+            cObj =>
+              (Number.isInteger(c) ? c : c.wordpress_id) === cObj.wordpress_id
+          ).id
       )
       delete e.categories
     }
@@ -490,6 +503,7 @@ exports.downloadMediaFiles = async ({
               cache,
               createNode,
               createNodeId,
+              parentNodeId: e.id,
               auth: _auth,
             })
 
@@ -607,3 +621,16 @@ exports.createNodesFromEntities = ({ entities, createNode }) => {
     })
   })
 }
+
+exports.createUrlPathsFromLinks = entities =>
+  entities.map(e => {
+    if (e.link && !e.path) {
+      try {
+        const link = new URL(e.link)
+        e.path = link.pathname
+      } catch (error) {
+        e.path = e.link
+      }
+    }
+    return e
+  })