+++ packages/gatsby/src/internal-plugins/webpack-theme-component-shadowing/index.js
@@ -35,14 +35,13 @@ module.exports = class GatsbyThemeComponentShadowingResolverPlugin {
       // get the location of the component relative to src/
       const [, component] = request.path.split(path.join(theme, `src`))
 
-      const builtComponentPath = this.resolveComponentPath({
-        matchingTheme: theme,
-        themes: this.themes,
-        component,
-        projectRoot: this.projectRoot,
-      })
+      const issuerExtension = path.extname(request.context.issuer)
 
-      if (!builtComponentPath) {
+      if (
+        request.context.issuer
+          .slice(0, -issuerExtension.length)
+          .endsWith(component)
+      ) {
         return resolver.doResolve(
           `describedRelative`,
           request,
@@ -51,14 +50,20 @@ module.exports = class GatsbyThemeComponentShadowingResolverPlugin {
           callback
         )
       }
-      const resolvedComponentPath = require.resolve(builtComponentPath)
-      // this callbackends the resolver fallthrough chain.
-      return callback(null, {
-        directory: request.directory,
-        path: resolvedComponentPath,
-        query: request.query,
-        request: ``,
+      const builtComponentPath = this.resolveComponentPath({
+        matchingTheme: theme,
+        themes: this.themes,
+        component,
+        projectRoot: this.projectRoot,
       })
+
+      return resolver.doResolve(
+        `describedRelative`,
+        { ...request, path: builtComponentPath || request.path },
+        null,
+        {},
+        callback
+      )
     })
   }
 
@@ -76,9 +81,11 @@ module.exports = class GatsbyThemeComponentShadowingResolverPlugin {
         path.join(path.resolve(`.`), `src`, theme),
       ]
         .concat(
-          themes.map(aTheme =>
-            path.join(path.dirname(require.resolve(aTheme)), `src`, theme)
-          )
+          Array.from(themes)
+            .reverse()
+            .map(aTheme =>
+              path.join(path.dirname(require.resolve(aTheme)), `src`, theme)
+            )
         )
         .map(dir => path.join(dir, component))
         .find(possibleComponentPath => {
@@ -97,7 +104,12 @@ module.exports = class GatsbyThemeComponentShadowingResolverPlugin {
               const filenameWithoutExtension = path.basename(filepath, ext)
               return filenameWithoutExtension
             })
-            .includes(path.basename(possibleComponentPath))
+            .includes(
+              path.basename(
+                possibleComponentPath,
+                path.extname(possibleComponentPath)
+              )
+            )
           return exists
         })
     }