+++ packages/gatsby/src/schema/__tests__/run-query.js
@@ -1,6 +1,4 @@
-const { GraphQLObjectType } = require(`graphql`)
 const nodesQuery = require(`../../db/nodes-query`)
-const { inferObjectStructureFromNodes } = require(`../infer-graphql-type`)
 const { store } = require(`../../redux`)
 require(`../../db/__tests__/fixtures/ensure-loki`)()
 
@@ -51,6 +49,7 @@ const makeNodes = () => [
     anArray: [1, 2, 5, 4],
     waxOnly: {
       foo: true,
+      bar: { baz: true },
     },
     anotherKey: {
       withANested: {
@@ -136,13 +135,19 @@ const makeNodes = () => [
 ]
 
 function makeGqlType(nodes) {
-  return new GraphQLObjectType({
-    name: `Test`,
-    fields: inferObjectStructureFromNodes({
-      nodes,
-      types: [{ name: `Test` }],
-    }),
+  const { createSchemaComposer } = require(`../../schema/schema-composer`)
+  const { addInferredFields } = require(`../infer/add-inferred-fields`)
+  const { getExampleValue } = require(`../infer/example-value`)
+
+  const sc = createSchemaComposer()
+  const typeName = `Test`
+  const tc = sc.createTC(typeName)
+  addInferredFields({
+    schemaComposer: sc,
+    typeComposer: tc,
+    exampleValue: getExampleValue({ nodes, typeName }),
   })
+  return tc.getType()
 }
 
 function resetDb(nodes) {
@@ -156,10 +161,8 @@ async function runQuery(queryArgs) {
   const nodes = makeNodes()
   resetDb(nodes)
   const gqlType = makeGqlType(nodes)
-  const context = {}
   const args = {
     gqlType,
-    context,
     firstOnly: false,
     queryArgs,
   }
@@ -199,8 +202,22 @@ describe(`Filter fields`, () => {
     expect(result[0].hair).toEqual(1)
   })
 
+  it(`handles ne: true operator`, async () => {
+    let result = await runFilter({ boolean: { ne: true } })
+
+    expect(result.length).toEqual(2)
+  })
+
   it(`handles nested ne: true operator`, async () => {
     let result = await runFilter({ waxOnly: { foo: { ne: true } } })
+
+    expect(result.length).toEqual(2)
+  })
+
+  it(`handles deeply nested ne: true operator`, async () => {
+    let result = await runFilter({
+      waxOnly: { bar: { baz: { ne: true } } },
+    })
 
     expect(result.length).toEqual(2)
   })
@@ -400,10 +417,7 @@ describe(`Filter fields`, () => {
     let result = await runFilter({ boolean: { nin: [true, null] } })
 
     expect(result.length).toEqual(1)
-    result.forEach(edge => {
-      expect(edge.boolean).not.toEqual(null)
-      expect(edge.boolean).not.toEqual(true)
-    })
+    expect(result[0].boolean).toBe(false)
   })
 
   it(`handles the glob operator`, async () => {
@@ -420,6 +434,20 @@ describe(`Filter fields`, () => {
     expect(result[0].index).toEqual(0)
     expect(result[1].index).toEqual(2)
   })
+
+  it(`handles the eq operator for array field values`, async () => {
+    const result = await runFilter({ anArray: { eq: 5 } })
+
+    expect(result.length).toBe(1)
+    expect(result[0].index).toBe(1)
+  })
+
+  it(`handles the ne operator for array field values`, async () => {
+    const result = await runFilter({ anArray: { ne: 1 } })
+
+    expect(result.length).toBe(1)
+    expect(result[0].index).toBe(2)
+  })
 })
 
 describe(`collection fields`, () => {
@@ -427,7 +455,7 @@ describe(`collection fields`, () => {
     let result = await runQuery({
       limit: 10,
       sort: {
-        fields: [`frontmatter___blue`],
+        fields: [`frontmatter.blue`],
         order: [`desc`],
       },
     })
@@ -470,7 +498,7 @@ describe(`collection fields`, () => {
     let result = await runQuery({
       limit: 10,
       sort: {
-        fields: [`frontmatter___blue`, `id`],
+        fields: [`frontmatter.blue`, `id`],
         order: [`desc`], // `id` field will be sorted asc
       },
     })
@@ -485,7 +513,7 @@ describe(`collection fields`, () => {
     let result = await runQuery({
       limit: 10,
       sort: {
-        fields: [`frontmatter___blue`, `id`],
+        fields: [`frontmatter.blue`, `id`],
         order: [`desc`, `desc`], // `id` field will be sorted desc
       },
     })