+++ themes/gatsby-theme-blog-mdx/src/components/mdx-theme-provider.js
@@ -1,0 +1,12 @@
+import React from "react"
+import { MDXProvider } from "@mdx-js/tag"
+import { TypographyProvider } from "typography-system"
+
+import theme from "../tokens"
+import mdxComponents from "./mdx"
+
+export default ({ children }) => (
+  <TypographyProvider theme={theme}>
+    <MDXProvider theme={mdxComponents}>{children}</MDXProvider>
+  </TypographyProvider>
+)