+++ packages/gatsby-plugin-google-gtag/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.16](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.0.15...gatsby-plugin-google-gtag@1.0.16) (2019-03-16)
+
+### Features
+
+- **gatsby-plugin-google-gtag:** add TS types for OutboundLink ([#12606](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/issues/12606)) ([bc72163](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/commit/bc72163)), closes [/github.com/gatsbyjs/gatsby/blob/3834b2bce5edc60cd5386561498ca29f028c1d30/packages/gatsby-plugin-google-analytics/src/index.js#L10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/issues/L10)
+
+## [1.0.15](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.0.14...gatsby-plugin-google-gtag@1.0.15) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-plugin-google-gtag
+
+## [1.0.14](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.0.13...gatsby-plugin-google-gtag@1.0.14) (2019-02-25)
+
+### Bug Fixes
+
+- **gatsby-plugin-google-gtag:** page view and title mismatch ([#12033](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/issues/12033)) ([b95ae69](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/commit/b95ae69))
+
+## [1.0.13](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.0.12...gatsby-plugin-google-gtag@1.0.13) (2019-02-04)
+
+### Bug Fixes
+
+- **gatsby-plugin-google-grag:** add additional check to prevent crash in Safari ([#11431](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/issues/11431)) ([#11503](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/issues/11503)) ([c79bc6d](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/commit/c79bc6d))
+
 ## [1.0.12](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.0.11...gatsby-plugin-google-gtag@1.0.12) (2019-02-01)
 
 **Note:** Version bump only for package gatsby-plugin-google-gtag